+++ cgbeacon2/server/blueprints/api_v1/views.py
@@ -9,10 +9,12 @@ from flask import (
     flash,
     redirect,
 )
+from flask_negotiate import consumes
 from cgbeacon2.constants import CHROMOSOMES
 from cgbeacon2.models import Beacon
 from cgbeacon2.utils.auth import authlevel
-from .controllers import create_allele_query, dispatch_query
+from cgbeacon2.utils.parse import validate_add_request
+from .controllers import create_allele_query, dispatch_query, add_variants
 
 API_VERSION = "1.0.0"
 LOG = logging.getLogger(__name__)
@@ -45,10 +47,10 @@ def info():
 @api1_bp.route("/apiv1.0/query_form", methods=["GET", "POST"])
 def query_form():
     """The endpoint to a super simple query form to interrogate this beacon
-       Query is performed only on public access datasets contained in this beacon
+    Query is performed only on public access datasets contained in this beacon
 
-       query_form page is accessible from a browser at this address:
-       http://127.0.0.1:5000/apiv1.0/query_form
+    query_form page is accessible from a browser at this address:
+    http://127.0.0.1:5000/apiv1.0/query_form
     """
 
     all_dsets = current_app.db["dataset"].find()
@@ -64,13 +66,9 @@ def query_form():
 
         else:  # query database
             # query database (it should return a datasetAlleleResponses object)
-            response_type = resp_obj["allelRequest"].get(
-                "includeDatasetResponses", "NONE"
-            )
+            response_type = resp_obj["allelRequest"].get("includeDatasetResponses", "NONE")
             query_datasets = resp_obj["allelRequest"].get("datasetIds", [])
-            exists, ds_allele_responses = dispatch_query(
-                query, response_type, query_datasets
-            )
+            exists, ds_allele_responses = dispatch_query(query, response_type, query_datasets)
             resp_obj["exists"] = exists
             resp_obj["error"] = None
             resp_obj["datasetAlleleResponses"] = ds_allele_responses
@@ -96,6 +94,32 @@ def query_form():
         dsets=all_dsets,
         form=dict(request.form),
     )
+
+
+@consumes("application/json")
+@api1_bp.route("/apiv1.0/add", methods=["POST"])
+def add():
+    """Endpoint accepting json data from POST requests. Save variants from a VCF file according to params specified in the request.
+    Example:
+    ########### POST request ###########
+    curl -X POST \
+    -H 'Content-Type: application/json' \
+    -d '{"dataset_id": "test_public",
+    "vcf_path": "path/to/cgbeacon2/resources/demo/test_trio.vcf.gz",
+    "samples" : ["ADM1059A1", "ADM1059A2"],
+    "assemblyId": "GRCh37"}' http://localhost:5000/apiv1.0/add
+    """
+
+    resp = None
+    # validate request content:
+    validate_request = validate_add_request(request)
+    if isinstance(validate_request, dict):  # If validation failed, return error response
+        resp = jsonify(validate_request)
+        resp.status_code = 422  # Unprocessable Entity
+        return resp
+    # Validation of request is OK, load eventual variants to db
+    resp = add_variants(request)
+    return resp
 
 
 @api1_bp.route("/apiv1.0/query", methods=["GET", "POST"])
@@ -129,9 +153,7 @@ def query():
     # Public access only has auth_levels = ([], False)
     auth_levels = authlevel(request, current_app.config.get("ELIXIR_OAUTH2"))
 
-    if isinstance(
-        auth_levels, dict
-    ):  # an error must have occurred, otherwise it's a tuple
+    if isinstance(auth_levels, dict):  # an error must have occurred, otherwise it's a tuple
         resp = jsonify(auth_levels)
         resp.status_code = auth_levels.get("errorCode", 403)
         return resp