+++ tests/server/blueprints/api_v1/test_request_errors.py
@@ -7,11 +7,116 @@ from cgbeacon2.constants import (
     INVALID_COORDINATES,
     BUILD_MISMATCH,
 )
+from cgbeacon2.resources import test_snv_vcf_path
 
 HEADERS = {"Content-type": "application/json", "Accept": "application/json"}
 
 BASE_ARGS = "query?assemblyId=GRCh37&referenceName=1&referenceBases=TA"
 ################## TESTS FOR HANDLING WRONG REQUESTS ################
+
+
+def test_add_no_dataset(mock_app):
+    """Test receiving a variant add request missing one of the required params"""
+    data = dict(vcf_path="path/to/vcf", assemblyId="GRCh37")
+    # When a POST add request is missing dataset id param:
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error 422 (Unprocessable Entity)
+    assert response.status_code == 422
+    resp_data = json.loads(response.data)
+    assert resp_data["message"] == "'dataset_id' is a required property"
+
+
+def test_add_no_vcf_path(mock_app):
+    """Test receiving a variant add request missing one of the required params"""
+    data = dict(dataset_id="test_id", assemblyId="GRCh37")
+    # When a POST add request is missing dataset path to vcf file
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error 422 (Unprocessable Entity)
+    assert response.status_code == 422
+    resp_data = json.loads(response.data)
+    assert resp_data["message"] == "'vcf_path' is a required property"
+
+
+def test_add_wrong_assembly(mock_app):
+    """Test receiving a variant add request with non-valid genome assembly"""
+    data = dict(dataset_id="test_id", vcf_path="path/to/vcf", assemblyId="FOO")
+    # When a POST add request is sent with a non valid assembly id
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error 422 (Unprocessable Entity)
+    assert response.status_code == 422
+    resp_data = json.loads(response.data)
+    assert resp_data["message"] == "'FOO' is not one of ['GRCh37', 'GRCh38']"
+
+
+def test_add_wrong_dataset(mock_app):
+    """Test receiving a variant add request with non-valid dataset id"""
+    data = dict(dataset_id="FOO", vcf_path="path/to/vcf", assemblyId="GRCh37")
+    # When a POST add request is sent with a non valid assembly id
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error
+    assert response.status_code == 422
+    # With message that dataset could not be found
+    data = json.loads(response.data)
+    assert data["message"] == "Provided dataset 'FOO' was not found on the server"
+
+
+def test_add_invalid_vcf_path(mock_app, public_dataset, database):
+    """Test receiving a variant add request with non-valid vcf path"""
+
+    # GIVEN a database containing a public dataset
+    database["dataset"].insert_one(public_dataset)
+
+    data = dict(dataset_id=public_dataset["_id"], vcf_path="path/to/vcf", assemblyId="GRCh37")
+    # When a POST add request is sent with a non valid assembly id
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error
+    assert response.status_code == 422
+    # With message that VCF path is not valid
+    data = json.loads(response.data)
+    assert data["message"] == "Error extracting info from VCF file, please check path to VCF"
+
+
+def test_add_invalid_samples(mock_app, public_dataset, database):
+    """Test receiving a variant add request with non-valid samples (samples not in provided VCF file)"""
+
+    # GIVEN a database containing a public dataset
+    database["dataset"].insert_one(public_dataset)
+
+    data = dict(
+        dataset_id=public_dataset["_id"],
+        vcf_path=test_snv_vcf_path,
+        assemblyId="GRCh37",
+        samples=["FOO", "BAR"],
+    )
+    # When a POST add request is sent with non-valid samples
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error
+    assert response.status_code == 422
+    # With message that VCF files doesn't contain those samples
+    data = json.loads(response.data)
+    assert "One or more provided samples were not found in VCF" in data["message"]
+
+
+def test_add_invalid_gene_list(mock_app, public_dataset, database):
+    """Test receiving a variant add request with non-valid genes object"""
+
+    # GIVEN a database containing a public dataset
+    database["dataset"].insert_one(public_dataset)
+
+    data = dict(
+        dataset_id=public_dataset["_id"],
+        vcf_path=test_snv_vcf_path,
+        assemblyId="GRCh37",
+        samples=["ADM1059A1"],
+        genes={"ids": [17284]},
+    )
+    # When a POST add request is sent with non-valid genes object (missing id_type for instance)
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    # Then it should return error
+    assert response.status_code == 422
+    # With message that missing info should be provided
+    data = json.loads(response.data)
+    assert "Please provide id_type (HGNC or Ensembl) for the given list of genes" in data["message"]
 
 
 def test_post_empty_query(mock_app):
@@ -52,9 +157,7 @@ def test_query_get_request_build_mismatch(mock_app, public_dataset):
 
     # When a request with genome build GRCh37 and detasetIds with genome build GRCh38 is sent to the server:
     query_string = "&".join([BASE_ARGS, f"datasetIds={public_dataset['_id']}"])
-    response = mock_app.test_client().get(
-        "".join(["/apiv1.0/", query_string]), headers=HEADERS
-    )
+    response = mock_app.test_client().get("".join(["/apiv1.0/", query_string]), headers=HEADERS)
 
     # Then it should return error
     assert response.status_code == 400
@@ -68,9 +171,7 @@ def test_query_get_request_missing_secondary_params(mock_app):
     """
     # When a request missing alternateBases or variantType params is sent to the server
     query_string = BASE_ARGS
-    response = mock_app.test_client().get(
-        "".join(["/apiv1.0/", query_string]), headers=HEADERS
-    )
+    response = mock_app.test_client().get("".join(["/apiv1.0/", query_string]), headers=HEADERS)
 
     # Then it should return error
     assert response.status_code == 400
@@ -83,9 +184,7 @@ def test_query_get_request_non_numerical_sv_coordinates(mock_app):
 
     query_string = "&".join([BASE_ARGS, "start=FOO&end=70600&variantType=DUP"])
     # When a request has a non-numerical start or stop position
-    response = mock_app.test_client().get(
-        "".join(["/apiv1.0/", query_string]), headers=HEADERS
-    )
+    response = mock_app.test_client().get("".join(["/apiv1.0/", query_string]), headers=HEADERS)
     data = json.loads(response.data)
     # Then it should return error
     assert response.status_code == 400
@@ -94,14 +193,12 @@ def test_query_get_request_non_numerical_sv_coordinates(mock_app):
 
 def test_query_get_request_missing_positions_params(mock_app):
     """Test the query endpoint by sending a request missing coordinate params:
-        Either start or any range coordinate
+    Either start or any range coordinate
 
     """
     # When a request missing start position and all the 4 range position coordinates (startMin, startMax, endMin, endMax)
     query_string = "&".join([BASE_ARGS, "alternateBases=T"])
-    response = mock_app.test_client().get(
-        "".join(["/apiv1.0/", query_string]), headers=HEADERS
-    )
+    response = mock_app.test_client().get("".join(["/apiv1.0/", query_string]), headers=HEADERS)
     data = json.loads(response.data)
     # Then it should return error
     assert response.status_code == 400
@@ -115,9 +212,7 @@ def test_query_get_request_non_numerical_range_coordinates(mock_app):
     query_string = "&".join([BASE_ARGS, range_coords])
 
     # When a request for range coordinates doesn't contain integers
-    response = mock_app.test_client().get(
-        "".join(["/apiv1.0/", query_string]), headers=HEADERS
-    )
+    response = mock_app.test_client().get("".join(["/apiv1.0/", query_string]), headers=HEADERS)
     data = json.loads(response.data)
     # Then it should return error
     assert response.status_code == 400